+++ src/auth0_token_error$_b.js
@@ -2,7 +2,7 @@ import { be_ } from '@ctx-core/object'
 import { atom$ } from '@ctx-core/nanostores'
 const key = 'auth0_token_error$'
 /** @type {import('auth0_token_error$_b.d.ts').auth0_token_error$_b} */
-export const auth0_token_error$_b = be_(key, (_ctx)=>{
+export const auth0_token_error$_b = be_(key, ()=>{
 	const auth0_token_error$ = atom$(undefined)
 	return auth0_token_error$
 })