+++ src/auth0_token_json$_b.js
@@ -3,7 +3,7 @@ import { be_ } from '@ctx-core/object'
 import { atom$ } from '@ctx-core/nanostores'
 const key = 'auth0_token_json$'
 /** @type {import('./auth0_token_json$_b.d.ts').auth0_token_json$_b} */
-export const auth0_token_json$_b = be_(key, (_ctx)=>{
+export const auth0_token_json$_b = be_(key, ()=>{
 	const auth_token_json$ = atom$(has_dom && localStorage.getItem(key) || null)
 	return auth_token_json$
 })