+++ src/get_jwks_json_b.js
@@ -1,4 +1,4 @@
-import { safe_fetch as fetch } from '@ctx-core/fetch-undici'
+import { fetch } from '@ctx-core/fetch-undici'
 import { be_ } from '@ctx-core/object'
 import { AUTH0_DOMAIN$_b } from './AUTH0_DOMAIN$_b.js'
 const key = 'get_jwks_json'
@@ -7,6 +7,7 @@ export const get_jwks_json_b = be_(key, (ctx)=>{
 	const AUTH0_DOMAIN$ = AUTH0_DOMAIN$_b(ctx)
 	return get_jwks_json
 	async function get_jwks_json() {
-		return fetch(`https://${AUTH0_DOMAIN$.$}/.well-known/jwks.json`)
+		const res = await fetch(`https://${AUTH0_DOMAIN$.$}/.well-known/jwks.json`)
+		return [await res.json, res]
 	}
 })