+++ src/logout_auth0_token_error_b.d.ts
@@ -1,4 +1,4 @@
 import { B } from '@ctx-core/object'
-import { auth0_token_error_T } from './auth0_token_error$_b.js'
+import type { Auth0Error } from 'auth0-js'
 export declare const logout_auth0_token_error_b:B<logout_auth0_token_error_T>
-export declare type logout_auth0_token_error_T = (error:auth0_token_error_T|null)=>void;
+export declare type logout_auth0_token_error_T = (error:Auth0Error|{ email:string }|null)=>void;