+++ src/post_auth0_dbconnections_signup_b.d.ts
@@ -1,4 +1,6 @@
 import { B } from '@ctx-core/object'
+import type { UserData } from 'auth0'
+import type { Auth0Error } from 'auth0-js'
 import type { auth0_client_id_body_I } from './_types'
 export declare const post_auth0_dbconnections_signup_b:B<post_auth0_dbconnections_signup_T>
 export interface post_auth0_dbconnections_signup_body_T extends auth0_client_id_body_I {
@@ -9,4 +11,5 @@ export interface post_auth0_dbconnections_signup_body_T extends auth0_client_id_
 	email:string;
 	password:string;
 }
-export declare type post_auth0_dbconnections_signup_T = (body:post_auth0_dbconnections_signup_body_T)=>Promise<Response>;
+export declare type post_auth0_dbconnections_signup_T =
+	(body:post_auth0_dbconnections_signup_body_T)=>Promise<[UserData|Auth0Error, Response]>