+++ src/validate_auth0_password_confirmation.d.ts
@@ -1,6 +1,6 @@
 import type { auth0_password_confirmation_data_I } from './_types'
-import type { auth0_token_error_I } from './auth0_token_error$_b.js'
+import type { Auth0Error } from 'auth0-js'
 export declare function validate_auth0_password_confirmation(data:auth0_password_confirmation_data_I):false|change_password_error_I|undefined;
-export interface change_password_error_I extends auth0_token_error_I {
+export interface change_password_error_I extends Auth0Error {
 	password_confirmation?:string;
 }